Implement queue listener

+++ src/support/sqs_listener.rs
@@ -2,28 +2,28 @@ use async_trait::async_trait;
 use aws_sdk_sqs::Client;
 use std::collections::HashMap;
 use std::sync::Arc;
-use tokio::task::JoinHandle;
+use tokio::task::JoinSet;
 use tokio::time::sleep;
 use tokio::time::Duration;
 
 pub struct SqsListenerBuilder {
-    client: Client,
+    client: Arc<Client>,
     listeners: HashMap<String, Arc<dyn SqsListener + Sync + Send>>,
-    delay: Duration,
+    delay: u64,
 }
 
 #[async_trait]
 pub trait SqsListener {
-    async fn on_message_received(&self, message: String) -> ();
+    async fn on_message_received(&self, message: String) -> Result<(), ()>;
 }
 
 
 impl SqsListenerBuilder {
-    pub async fn from(client: Client) -> SqsListenerBuilder {
-        SqsListenerBuilder { client, listeners: HashMap::new(), delay: Duration::from_millis(300) }
+    pub async fn from(client: Arc<Client>) -> SqsListenerBuilder {
+        SqsListenerBuilder { client, listeners: HashMap::new(), delay: 300 }
     }
 
-    pub fn polling_delay(mut self, delay: Duration) -> SqsListenerBuilder {
+    pub fn polling_delay(mut self, delay: u64) -> SqsListenerBuilder {
         self.delay = delay;
         self
     }
@@ -33,37 +33,46 @@ impl SqsListenerBuilder {
         self
     }
 
-    pub fn run(self) -> JoinHandle<()> {
-        tokio::task::spawn(async move {
-            let listeners = self.listeners;
-            let client = self.client;
-            let delay = self.delay;
-            loop {
-                for (queue_url, consumer) in &listeners {
-                    SqsListenerBuilder::receive(&client, queue_url, &consumer).await;
+    pub async fn run(self) {
+        let listeners = self.listeners;
+        let client = self.client;
+        let delay = self.delay;
+
+        let mut join_set = JoinSet::new();
+
+        for (queue_url, consumer) in listeners {
+            let client_clone = Arc::clone(&client);
+            join_set.spawn(async move {
+                loop {
+                    receive_queue_message(&client_clone, &queue_url, &consumer).await;
+                    sleep(Duration::from_millis(delay)).await;
                 }
-                sleep(delay).await;
-            }
-        })
+            });
+        }
+
+        while let Some(_) = join_set.join_next().await {}
     }
+}
 
-    async fn receive(client: &Client, queue_url: &String, consumer: &Arc<dyn SqsListener + Sync + Send>) {
-        let rcv_message_output = client.receive_message()
-            .queue_url(queue_url)
-            .send().await.unwrap();
+async fn receive_queue_message(
+    client: &Client,
+    queue_url: &String,
+    consumer: &Arc<dyn SqsListener + Sync + Send>,
+) {
+    let rcv_message_output = client.receive_message()
+        .queue_url(queue_url)
+        .send().await.unwrap();
 
-        for message in rcv_message_output.messages.unwrap_or_default() {
-            let body = message.body.unwrap_or_default();
+    for message in rcv_message_output.messages.unwrap_or_default() {
+        let body = message.body.unwrap_or_default();
 
-            tracing::info!("Received message from queue '{}' : {}", queue_url, body);
-
-            consumer.on_message_received(body).await;
-
+        if consumer.on_message_received(body).await.is_ok() {
             client.delete_message()
                 .queue_url(queue_url)
                 .receipt_handle(message.receipt_handle.unwrap())
                 .send()
-                .await.unwrap();
+                .await
+                .unwrap();
         }
     }
 }
